refactor(features): move app-header and app-sidebar-header out of components
